SBERDOMA-503 review fixes - move ArrowIcons component to icons dir & extract const for header action
